appkit canvas: move key character decoding into Elisa

+++ test/appkit_canvas_bridge_test.m
@@ -83,10 +83,10 @@ void elisa_appkit_canvas_pointer_scroll(float x, float y, float dx, float dy) {
     test_pointer_button = 0;
 }
 void elisa_appkit_canvas_raw_key(int down, int keyCode, int character) { (void)down; (void)keyCode; (void)character; }
-void elisa_appkit_canvas_key_down_event(size_t event, int keyCode, int character, size_t modifiers) {
+void elisa_appkit_canvas_key_down_event(size_t event, int keyCode, size_t character, size_t modifiers) {
     (void)event; (void)keyCode; (void)character; (void)modifiers;
 }
-void elisa_appkit_canvas_key_up(int keyCode, int character) { (void)keyCode; (void)character; }
+void elisa_appkit_canvas_key_up(int keyCode, size_t character) { (void)keyCode; (void)character; }
 void elisa_appkit_canvas_raw_flags(int keyCode, size_t modifiers) {
     (void)keyCode; (void)modifiers;
 }
